Update and document the build process
Current process would not work when using locally with apollo-server
and it surfaced a lot of problems. These fixes _should_ solve them.

Also, TIL that you cannot use npm link across different directories
where one is the source and other is dependent when you have a peer
dependency in your source directory. This only work if the dependent
directory is nested somewhere in the source directory structure. This
is because when you use npm link in the depedent directory and then
node or your bundler tries to locate the peer dependency, the look up
algorithm starts off in the source directory since that is where your
"main" bundle is sitting. Even if your _dependent_ tree _has_ the
peerDep, this look up can fail if you haven't installed the peerDep in
the source tree as well.

God that fucked me up for a bit.